refactor: grpc utils
