+++ internal/grpc_utils/common_api/interface.go
